clean code smells

+++ src/cpp/benders/benders_by_batch/include/BatchCollection.h
@@ -15,11 +15,11 @@ class BatchCollection {
 
  public:
   BatchCollection(const std::vector<std::string>& sub_problem_names,
-                  unsigned size_of_sub_problems_collection);
+                  size_t size_of_sub_problems_collection);
   class InvalidSizeOfSubPrblemCollection : public std::runtime_error {
    public:
-    explicit InvalidSizeOfSubPrblemCollection(unsigned size,
-                                              unsigned sub_problems_names_size)
+    explicit InvalidSizeOfSubPrblemCollection(size_t size,
+                                              size_t sub_problems_names_size)
         : std::runtime_error(std::string(
               "Bacth size: " + std::to_string(size) +
               " must be less or equal than number of sub problems: " +
